Put terrain generation codes into dedicated source file
